feat: fine-tune map on both platforms and clean up code
#45
