+++ composeApp/src/commonMain/kotlin/features/atmfinder/AtmFinderScreen.kt
@@ -5,29 +5,15 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Circle
-import androidx.compose.material.icons.filled.LocationOn
-import androidx.compose.material.icons.outlined.Accessible
-import androidx.compose.material.icons.outlined.AttachMoney
-import androidx.compose.material.icons.outlined.House
-import androidx.compose.material.icons.outlined.MoneyOff
-import androidx.compose.material.icons.outlined.NaturePeople
-import androidx.compose.material.icons.outlined.NotAccessible
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -37,8 +23,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import bankingapp.composeapp.generated.resources.Res
-import bankingapp.composeapp.generated.resources.atm_finder_default_atm_name
-import bankingapp.composeapp.generated.resources.atm_finder_distance_meters
+import bankingapp.composeapp.generated.resources.atm_finder_loading_loading_atms
+import bankingapp.composeapp.generated.resources.atm_finder_loading_waiting_for_location
+import bankingapp.composeapp.generated.resources.atm_finder_loading_waiting_for_permission
 import bankingapp.composeapp.generated.resources.atm_finder_title
 import components.BackButton
 import components.DefaultPosition
@@ -55,8 +42,6 @@ import theme.AppTheme
 import theme.dp16
 import theme.dp24
 import theme.dp8
-import kotlin.math.roundToInt
-
 
 @Composable
 fun AtmFinderScreen(
@@ -67,18 +52,21 @@ fun AtmFinderScreen(
     }
     val location = viewModel.location.collectAsState(null).value
     val loadingState by viewModel.loadingState.collectAsState()
-    LaunchedEffect(loadingState) {
-        println("`````ls: $loadingState")
-    }
 
     bindPermissionsController(viewModel.permissionsController)
     bindLocationTracker(viewModel.locationTracker)
 
     when (loadingState) {
         LoadingState.Initial -> FullScreenLoading()
-        LoadingState.WaitingForPermission -> FullScreenLoading("waiting for permission")
-        LoadingState.WaitingForLocation -> FullScreenLoading("waiting for location")
-        LoadingState.LoadingAtms -> FullScreenLoading("loading atms")
+        LoadingState.WaitingForPermission ->
+            FullScreenLoading(stringResource(Res.string.atm_finder_loading_waiting_for_permission))
+
+        LoadingState.WaitingForLocation ->
+            FullScreenLoading(stringResource(Res.string.atm_finder_loading_waiting_for_location))
+
+        LoadingState.LoadingAtms ->
+            FullScreenLoading(stringResource(Res.string.atm_finder_loading_loading_atms))
+
         LoadingState.Success ->
             AtmFinderScreenContent(
                 location = location,
@@ -143,7 +131,7 @@ fun AtmFinderScreenContent(
                     item {
                         LocationVisualizer(
                             position = location ?: DefaultPosition,
-                            modifier = Modifier.fillMaxWidth().aspectRatio(2f),
+                            modifier = Modifier.fillMaxWidth().aspectRatio(1.5f),
                             markers = atms.take(10).map { atm ->
                                 Marker(
                                     position = GpsPosition(atm.lat, atm.lon),
@@ -154,148 +142,33 @@ fun AtmFinderScreenContent(
                             parentScrollEnableState = remember { mutableStateOf(false) },
                         )
                     }
-                    atms.forEachIndexed { index, atm ->
-                        item {
-                            Box(
-                                modifier = Modifier
-                                    .background(AppTheme.colors.backgroundNeutral)
-                                    .padding(horizontal = dp16)
-                                    .padding(
-                                        top = if (index == atms.indices.first) dp16 else dp8,
-                                        bottom = if (index == atms.indices.last) dp16 else dp8
+                    atms.sortedBy { atm ->
+                        location?.distanceToInMeters(atm.getLocation())
+                    }
+                        .forEachIndexed { index, atm ->
+                            item {
+                                Box(
+                                    modifier = Modifier
+                                        .background(AppTheme.colors.backgroundNeutral)
+                                        .padding(horizontal = dp16)
+                                        .padding(
+                                            top = if (index == atms.indices.first) dp16 else dp8,
+                                            bottom = if (index == atms.indices.last) dp16 else dp8
+                                        )
+                                ) {
+                                    Atm(
+                                        atm = atm,
+                                        index = index,
+                                        distanceInMeters = location?.distanceToInMeters(atm.getLocation()),
                                     )
-                            ) {
-                                Atm(
-                                    atm = atm,
-                                    index = index,
-                                    distanceInMeters = location?.distanceToInMeters(atm.getLocation()),
-                                )
+                                }
                             }
                         }
-                    }
                 }
             }
         }
     }
 }
 
-@Composable
-private fun Atm(
-    atm: Atm,
-    index: Int,
-    distanceInMeters: Double?,
-) {
-    Surface(
-        shadowElevation = dp8,
-        color = AppTheme.colors.surfaceNeutral,
-        shape = RoundedCornerShape(dp8),
-    ) {
-        Row(
-            modifier = Modifier.padding(dp16).fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(dp8),
-        ) {
-            AtmPinIcon(index = index)
-            Column {
-                Text(
-                    text = atm.name ?: stringResource(Res.string.atm_finder_default_atm_name),
-                    fontSize = 13.sp,
-                    color = AppTheme.colors.textDarker
-                )
-                atm.getFullAddress()?.let { fullAddress ->
-                    Text(
-                        text = fullAddress,
-                        fontSize = 10.sp,
-                        color = AppTheme.colors.textDarker,
-                    )
-                }
-                distanceInMeters?.let {
-                    Text(
-                        text = stringResource(Res.string.atm_finder_distance_meters, distanceInMeters.roundToInt()),
-                        fontSize = 10.sp,
-                        color = AppTheme.colors.textDarker,
-                    )
-                }
-                AtmFeatures(atm = atm)
-            }
-        }
-    }
-}
-
-@Composable
-private fun AtmPinIcon(
-    index: Int,
-) {
-    val isIndexed = (index + 1) in 1..5
-
-    Box(contentAlignment = Alignment.Center) {
-        Icon(
-            imageVector = if (isIndexed) {
-                Icons.Filled.Circle
-            } else {
-                Icons.Filled.LocationOn
-            },
-            tint = AppTheme.colors.textDarker,
-            contentDescription = null
-        )
-        if (isIndexed) {
-            Text(
-                text = (index + 1).toString(),
-                fontSize = 8.sp,
-                color = AppTheme.colors.backgroundNeutral
-            )
-        }
-    }
-}
-
-@Composable
-private fun AtmFeatures(
-    atm: Atm,
-) {
-    Row(
-        horizontalArrangement = Arrangement.spacedBy(dp8)
-    ) {
-        if (atm.wheelchair != null) {
-            Icon(
-                imageVector = if (atm.wheelchair) {
-                    Icons.Outlined.Accessible
-                } else {
-                    Icons.Outlined.NotAccessible
-                },
-                tint = AppTheme.colors.main,
-                contentDescription = null
-            )
-        }
-        if (atm.indoor != null) {
-            Icon(
-                imageVector = if (atm.indoor) {
-                    Icons.Outlined.House
-                } else {
-                    Icons.Outlined.NaturePeople
-                },
-                tint = AppTheme.colors.main,
-                contentDescription = null
-            )
-        }
-        if (atm.cashIn != null) {
-            Icon(
-                imageVector = if (atm.cashIn) {
-                    Icons.Outlined.AttachMoney
-                } else {
-                    Icons.Outlined.MoneyOff
-                },
-                tint = AppTheme.colors.main,
-                contentDescription = null
-            )
-        }
-
-    }
-}
-
-private fun Atm.getFullAddress(): String? =
-    if (postcode == null && city == null && street == null && houseNumber == null) {
-        null
-    } else {
-        listOfNotNull(postcode, city, street, houseNumber).joinToString(" ")
-    }
 
 expect fun getPermissionsController(): PermissionsController
